Try to resolve missing members in tmwxHtmlHelpFrame
Both m_Printer and m_HtmlWin are nowhere to be found. The printer seems
to come in via the setter, so we can probably just put it in our class.
Presumably it was previously in some base.

As for m_HtmlWin I hope I got the right one. A method GetOpenedPage
exists for exactly one class in wxWidgets: wxHtmlWindow. That seems to
be obtained by these calls. I can't test it yet, let's see.